Frames moved to separate folder and namespace
