+++ PerformanceMeter/PerformanceMeter/EndGameDisplay.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using MelonLoader;
+using PerformanceMeter.Frames;
 using UnityEngine;
 using UnityEngine.UI;
