+++ PerformanceMeter/PerformanceMeter/Frames/PercentFrame.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PerformanceMeter
+namespace PerformanceMeter.Frames
 {
     public class PercentFrame
     {
