+++ PerformanceMeter/PerformanceMeter/Utils.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using PerformanceMeter.Frames;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
